+++ src/main/java/com/besedin/skypro/calculator/service/CalculatorService.java
@@ -5,19 +5,19 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CalculatorService {
-    public int plus(int a, int b) {
+    public double plus(double a, double b) {
         return a + b;
     }
 
-    public int minus(int a, int b) {
+    public double minus(double a, double b) {
         return a - b;
     }
 
-    public int multiply(int a, int b) {
+    public double multiply(double a, double b) {
         return a * b;
     }
 
-    public double divide(int a, int b) {
+    public double divide(double a, double b) {
         if(b == 0){
             throw new DontDivideOnOException("Делить на ноль нельзя");
         }
